Created a reusable (abstract) Inventory System
The system consists of a generic Inventory class that stores the generic type and invokes events when an item is added or removed.

The UI gets updated by calling the UpdateUI method from the Interface. The rest of the UI is an example implementation.

The updating of the UI is done via the ItemInventoryManager.