Created a reusable character controller 2D & player movement 2D class
The character controller 2D holds the functionality for moving, jumping and crouching. The character controller supports both platformer style- as well as top-down style movement, by toggling the gravity boolean.

To test the character controller, I created a player movement class that handles the input and interfaces with the character controller to drive the player movement.