Created a reusable character controller 3D & player movement 3D class
The character controller 3D holds the functionality for moving, jumping, sprinting and crouching. The character controller supports both normal and floating style movement, by toggling the gravity boolean.

To test the character controller, I created a player movement class that handles the input and interfaces with the character controller to drive the player movement.